Added control in get_kegg_reference().
I spotted a problem in Kegg genome database, where the id: "gn:T07039" had a REFERENCE voice, but the reference ID was just a blank spot (even if it had a title and an author). This issue may very likely happen in other ids, so I added an additional control on the length of the string vector.
Note: In that particular case we can find the reference id by manually serching the title of the article, but I simply labeled it as a "None" entry.